Cleaning up Part 2
